backend/requirements.txt
flask
flask-cors
pocket-tts
soundfile
transformers
torch
torchvision
pillow
timm
einops
requests
python-dotenv
browser-use
langchain-google-genai
pymongo
